changed duck controller and duck actions methods
