--- src/test/java/autotests/DuckCreateTest.java
@@ -1,47 +1,0 @@
-package autotests;
-
-import com.consol.citrus.TestCaseRunner;
-import com.consol.citrus.annotations.CitrusResource;
-import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.message.MessageType;
-import com.consol.citrus.validation.json.JsonPathMessageValidationContext;
-import org.springframework.http.HttpStatus;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Test;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
-import static com.consol.citrus.validation.json.JsonPathMessageValidationContext.Builder.jsonPath;
-
-
-public class DuckCreateTest extends DuckBaseTest {
-    @Test(description = "Проверка создания утки с материалом rubber")
-    @CitrusTest
-    public void createDuckWithMaterialRubber(@Optional @CitrusResource TestCaseRunner runner){
-        createDuck(runner, "red", 0.53, "rubber", "quack", "FIXED");
-        validateDuckResponse(runner, "red", 0.53, "rubber", "quack", "FIXED");
-    }
-
-    @Test(description = "Проверка создания утки с материалом wood")
-    @CitrusTest
-    public void createDuckWithMaterialWood(@Optional @CitrusResource TestCaseRunner runner){
-        createDuck(runner, "red", 0.53, "wood", "quack", "FIXED");
-        validateDuckResponse(runner, "red", 0.53, "wood", "quack", "FIXED");
-    }
-
-    public void validateDuckResponse(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response(HttpStatus.OK)
-                        .message()
-                        .type(MessageType.JSON)
-                        .validate(jsonPath()
-                                .expression("$.color", color)
-                                .expression("$.height", height)
-                                .expression("$.material", material)
-                                .expression("$.sound", sound)
-                                .expression("$.wingsState", wingsState)
-                        )
-        );
-    }
-}
