+++ src/test/java/autotests/tests/duckActionController/DuckQuackTest.java
@@ -1,17 +1,15 @@
-package autotests;
+package autotests.tests.duckActionController;
 
+import autotests.clients.DuckActionsClient;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
 import java.util.concurrent.atomic.AtomicInteger;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
 
-public class DuckQuackTest extends DuckBaseTest {
+public class DuckQuackTest extends DuckActionsClient {
 
     @Test (description = "Проверка того, что утки с четным id издают звуки")
     @CitrusTest
@@ -26,7 +24,7 @@ public class DuckQuackTest extends DuckBaseTest {
                 id.set(Integer.parseInt(action.getVariable("duckId")));
             });
         } while (id.get() % 2 != 0);
-        checkDuckQuack(runner, String.valueOf(id.get()), 2, 3, "quack-quack-quack, quack-quack-quack");
+        duckQuack(runner, String.valueOf(id.get()), 2, 3, "quack-quack-quack, quack-quack-quack");
     }
 
     @Test (description = "Проверка того, что утки с нечетным id издают звуки")
@@ -42,23 +40,6 @@ public class DuckQuackTest extends DuckBaseTest {
                 id.set(Integer.parseInt(action.getVariable("duckId")));
             });
         } while (id.get() % 2 == 0);
-        checkDuckQuack(runner, String.valueOf(id.get()), 2, 3, "quack-quack-quack, quack-quack-quack");
-    }
-
-    private void checkDuckQuack(TestCaseRunner runner, String duckId, int repetitionCount, int soundCount, String expectedSound) {
-        runner.$(
-                http().client("http://localhost:2222").send().get("/api/duck/action/quack")
-                        .queryParam("id", duckId)
-                        .queryParam("repetitionCount", String.valueOf(repetitionCount))
-                        .queryParam("soundCount", String.valueOf(soundCount))
-        );
-
-        runner.$(
-                http().client("http://localhost:2222").receive().response(HttpStatus.OK)
-                        .message().contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body("{\n" +
-                                "\"sound\": \"" + expectedSound + "\"\n" +
-                                "}")
-        );
+        duckQuack(runner, String.valueOf(id.get()), 2, 3, "quack-quack-quack, quack-quack-quack");
     }
 }
