+++ src/test/java/autotests/tests/duckController/DuckDeleteTest.java
@@ -1,14 +1,14 @@
-package autotests;
+package autotests.tests.duckController;
 
+import autotests.clients.DuckActionsClient;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
-import org.springframework.http.HttpStatus;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
-public class DuckDeleteTest extends DuckBaseTest {
+
+public class DuckDeleteTest extends DuckActionsClient {
     @Test(description = "Проверка удаления утки")
     @CitrusTest
     public void successfulDeleteDuck(@Optional @CitrusResource TestCaseRunner runner){
@@ -16,15 +16,5 @@ public class DuckDeleteTest extends DuckBaseTest {
         saveDuckId(runner);
         deleteDuck(runner, "${duckId}");
         checkDuckDeleted(runner, "${duckId}");
-    }
-
-    private void checkDuckDeleted(TestCaseRunner runner, String duckId) {
-        runner.$(
-                http().client("http://localhost:2222").send().get("/api/duck/action/properties").queryParam("id", duckId)
-        );
-
-        runner.$(
-                http().client("http://localhost:2222").receive().response(HttpStatus.INTERNAL_SERVER_ERROR)
-        );
     }
 }
